Moved helper classes to hazenlib.utils.py
