+++ hazenlib/utils.py
@@ -6,7 +6,8 @@ import matplotlib
 import numpy as np
 
 from collections import defaultdict
-from skimage import filters
+from skimage import filters, measure
+from typing import Union
 
 import hazenlib.exceptions as exc
 from hazenlib.logger import logger
@@ -642,3 +643,192 @@ class ShapeDetector:
             size = (size[1], size[0])
             angle = angle - 90
             return (x, y), size, angle
+
+
+class Point:
+    """Class representing a point in Cartesian Space"""
+
+    def __init__(self, x: Union[int, float], y: Union[int, float]):
+        if not isinstance(x, (int, float)):
+            raise ValueError("arg x of Point.__init__ should be int or float.")
+        if not isinstance(y, (int, float)):
+            raise ValueError("arg y of Point.__init__ should be int or float.")
+        self._xy = np.array([x, y])
+
+    @property
+    def x(self):
+        """Getter for x coordinate."""
+        return self._xy[0]
+
+    @property
+    def y(self):
+        """Getter for y coord"""
+        return self._xy[1]
+
+    @property
+    def xy(self) -> np.ndarray:
+        """Getter for xy np array"""
+        return self._xy
+
+    def get_distance_to(self, other: "Point") -> float:
+        """Calculates distance between two point objects.
+
+        Args:
+            other (Point): Point to calculate distance to.
+
+        Returns:
+            dist (float): The distance between the points.
+        """
+        if not isinstance(other, type(self)):
+            raise ValueError("arg other of Point.get_distance_to should be Point.")
+        dist = np.sqrt((self.x - other.x) ** 2 + (self.y - other.y) ** 2)
+        return dist
+
+    def as_int(self) -> "Point":
+        """Returns an instance of Point with coords mapped to int.
+
+        Returns:
+            as_int (Point): Instance of Point with coords mapped to int.
+        """
+        as_int = type(self)(x=round(self.x), y=round(self.y))
+        return as_int
+
+    def __add__(self, other: "Point") -> "Point":
+        """Addition between two point objects
+
+        Args:
+            other (Point): Point to add.
+
+        Returns:
+            result (Point): Point object for summed coords
+        """
+        if not isinstance(other, type(self)):
+            raise ValueError("arg other of Point.__add__ should be Point")
+        result = type(self)(x=self.x + other.x, y=self.y + other.y)
+        return result
+
+    def __sub__(self, other: "Point") -> "Point":
+        """Subtraction between two point objects
+
+        Args:
+            other (Point): Point to subtract.
+
+        Returns:
+            result (Point): Point object for subtracted coords
+        """
+        if not isinstance(other, type(self)):
+            raise ValueError("arg other of Point.__sub__ should be Point")
+        result = type(self)(x=self.x - other.x, y=self.y - other.y)
+        return result
+
+    def __truediv__(self, scalar: Union[int, float]) -> "Point":
+        """Divides point by scalar.
+
+        Args:
+            scalar (int, float): scalar to divide point by.
+
+        Returns:
+            result (Point): Point object for divided coords
+        """
+        if not isinstance(scalar, (int, float)):
+            raise ValueError("arg scalar of Point.__truediv__ should be int or float.")
+        result = type(self)(x=self.x / scalar, y=self.y / scalar)
+        return result
+
+    def __mul__(self, scalar: Union[int, float]) -> "Point":
+        """Multiplies point by scalar
+
+        Args:
+            scalar (int, float): scalar to multiply point by.
+
+        Returns:
+            result (Point): Point object for multiplied coords
+        """
+        if not isinstance(scalar, (int, float)):
+            raise ValueError("arg scalar of Point.__mul__ should be int or float.")
+        result = type(self)(x=self.x * scalar, y=self.y * scalar)
+        return result
+
+    def __str__(self) -> str:
+        return f"Point{(self.x, self.y)}"
+
+
+class Line:
+    """Class for a line in Cartesian Space"""
+
+    def __init__(self, p1: "Point", p2: "Point"):
+        if not isinstance(p1, Point):
+            raise ValueError("arg p1 of Line.__init__ should be Point.")
+        if not isinstance(p2, Point):
+            raise ValueError("arg p2 of Line.__init__ should be Point.")
+
+        self._p1 = p1
+        self._p2 = p2
+
+    @property
+    def p1(self) -> "Point":
+        """Getter for point 1 in line"""
+        return self._p1
+
+    @property
+    def p2(self) -> "Point":
+        """Getter for point 2 in line"""
+        return self._p2
+
+    @property
+    def midpoint(self) -> "Point":
+        """Getter for line midpoint"""
+        return (self._p1 + self._p2) / 2
+
+    def get_profile(self, refImg: np.ndarray) -> list:
+        """Gets profile across line using pixel values from reference image
+
+        Args:
+            refImg (np.ndarray): Reference image for obtaining pixel values
+
+        Returns:
+            profile (list): pixel value profile across line
+        """
+        profile = measure.profile_line(
+            image=refImg,
+            src=self._p1.as_int().xy.tolist()[::-1],
+            dst=self._p2.as_int().xy.tolist()[::-1],
+        )
+        return profile
+
+    def point_swap(self):
+        """Swaps order of points"""
+        self._p1, self._p2 = self._p2, self._p1
+
+    def get_subline(self, percOfOrig: Union[int, float]) -> "Line":
+        """Returns an instance of self that is reduced in length to be X percent of the original. \n
+        The percentage for shrinking is determined by var percOfOrig
+
+        Args:
+            percOfOrig (int, float): Percentage of original line to shrink output line to
+
+        Returns:
+            subline (Line): subline of original line
+        """
+        if not isinstance(percOfOrig, (int, float)):
+            raise ValueError("arg percOfOrig of Line.get_subline should be int or float.")
+        if not 0 < percOfOrig <= 100:
+            raise ValueError(
+                f"For Line.get_subline method:\n\t   Arg perc should be in bounds (0, 100], received perc={percOfOrig}."
+            )
+        percOffSide = (100 - percOfOrig) / 2
+        vector = self._p1 - self._p2
+        p1Prime = self._p1 - vector * percOffSide / 100
+        p2Prime = self._p2 + vector * percOffSide / 100
+        subline = type(self)(p1=p1Prime, p2=p2Prime)
+
+        return subline
+
+    def __iter__(self) -> iter:
+        x_values = [self._p1.x, self._p2.x]
+        y_values = [self._p1.y, self._p2.y]
+        return iter([x_values, y_values])
+
+    def __str__(self) -> str:
+        """Str representation of Line"""
+        return f"Line(\n    p1={self._p1},\n    p2={self._p2}\n)"
